chore: split CI so contributors' deployments can work (#202)
This PR splits up deployments by moving PR deployments to a privileged
YAML file that uses an artifact. This is because PRs from forks don't
have access to secrets, and thus can't deploy.

The way it works is:
1. The normal CI runs, building and saving the HTML
2. Another workflow is triggered when the main one completes, and if it
saved HTML, labels the PR with `HTML available`
3. Adding this label triggers a privileged deployment workflow

The idea is that it should be easy to trigger a new deployment when
needed, and the status should be clearly associated with the PR. There's
no way to test it without merging, I think, so the plan is:

1. Merge while keeping the old PR deployment logic
2. Check that the new logic works
3. Delete the old logic

This is in case work must be paused due to other priorities - nothing
gets worse underway.